requirements.txt
requests>=2.31.0
PyYAML>=6.0
jsonschema>=4.17.0
python-dateutil>=2.8.2
pytest>=7.0.0
pytest-cov>=4.0.0
